Updated the German translation file, removed errors and spelling mistakes - german.js
Hey YouClipped team,
I'm new to GitHub, so sorry if I did something wrong here.
I stumbled over YouClipped thru Loogle+ and saw that you could help translate it. And so I thought I will help to translate it into German, as I'm a native German speaker and live in Germany. When I saw that there was already a complete translation file, I thought "Oh cool!" and took a look at it. 
But then I saw that there were spelling mistakes and stuff that was translated word for word and makes no sense like that.
So I made a GitHub Account and edited the file. Now there should be no errors in the file anymore. 
By the looks of it, you can see what exactly I have changed and I hope you will see this.
I sadly couldn't correct everything due to the way the translation code works (for example, the "Upload"-Button and the text "Uploads" on the home page use the same word in the file, but in German these are two diffrent words. If you use the same for both, it makes no sense. So you either have the text on the home page make no sense, or the upload button text.)
I also made it less formal by changing the pronouns from the formal version to the less and more informal version. I think it sounds nicer and more inviting like that.
Hopefully this gets implemented! 
Thanks for your work and making it possible to contribute by translating and stuff.
Have a nice weekend :)